+++ util-security/src/main/scala/com/twitter/util/security/X509CertificateFile.scala
@@ -3,8 +3,10 @@ package com.twitter.util.security
 import com.twitter.logging.Logger
 import com.twitter.util.Try
 import com.twitter.util.security.X509CertificateFile._
-import java.io.{ByteArrayInputStream, File}
-import java.security.cert.{CertificateFactory, X509Certificate}
+import java.io.File
+import java.nio.charset.StandardCharsets
+import java.nio.file.Files
+import java.security.cert.X509Certificate
 
 /**
  * A representation of an X.509 Certificate PEM-encoded and stored
@@ -23,38 +25,23 @@ class X509CertificateFile(file: File) {
   /**
    * Attempts to read the contents of the X.509 Certificate from the file.
    */
-  def readX509Certificate(): Try[X509Certificate] = {
-    val pemFile = new PemFile(file)
-    pemFile
-      .readMessage(MessageType)
-      .map(generateX509Certificate)
-      .onFailure(logException)
-  }
+  def readX509Certificate(): Try[X509Certificate] = Try {
+    new String(Files.readAllBytes(file.toPath), StandardCharsets.UTF_8)
+  }.flatMap { buffered =>
+      X509CertificateDeserializer.deserializeCertificate(buffered, file.getName)
+    }.onFailure(logException)
 
   /**
    * Attempts to read the contents of multiple X.509 Certificates from the file.
    */
-  def readX509Certificates(): Try[Seq[X509Certificate]] = {
-    val pemFile = new PemFile(file)
-    pemFile
-      .readMessages(MessageType)
-      .map(certBytes => certBytes.map(generateX509Certificate))
-      .onFailure(logException)
-  }
+  def readX509Certificates(): Try[Seq[X509Certificate]] = Try {
+    new String(Files.readAllBytes(file.toPath), StandardCharsets.UTF_8)
+  }.flatMap { buffered =>
+      X509CertificateDeserializer.deserializeCertificates(buffered, file.getName)
+    }.onFailure(logException)
 
 }
 
 private object X509CertificateFile {
-  private val MessageType: String = "CERTIFICATE"
-
   private val log = Logger.get("com.twitter.util.security")
-
-  private def generateX509Certificate(decodedMessage: Array[Byte]): X509Certificate = {
-    val certFactory = CertificateFactory.getInstance("X.509")
-    val certificate = certFactory
-      .generateCertificate(new ByteArrayInputStream(decodedMessage))
-      .asInstanceOf[X509Certificate]
-    certificate.checkValidity()
-    certificate
-  }
 }